Replace main code to the fsm
